test(storybook): DeleteAccountBanner DEV-1117 (#6338)
### 💭 Notes

Changes:
- Moved `meMock` and `environmentMock` to be globally defined for all
stories
- Created `TestIdSessionDecorator` that reads `args.storybookTestId` of
given story and stores it in sessionStorage making it accessible for MSW
mocks
- `DeleteAccountBanner.stories.tsx` now has two more stories: User Has
Assets, and User Owns MMO
- `DeleteAccountBanner.tsx` no longer needs test code to be present
- `assets.mocks.ts` and `organization.mocks.ts` use `storybookTestId`
from `sessionStorage` to serve custom responses
- Bumped `msw` and `msw-storybook-addon` versions while debugging this,
hoping it would solve an issue I had with msw returning 404s for
existing mocks. The problem solved by itself :shrug: but this doesn't
break anything so I left it here.

The problem was caused by mocks not returning proper customized
responses for non-default cases. I redid the way we handle those and was
able to do it in such a way that we no longer need to adjust original
components.

**Unfortunately** the solution above will not work for "Docs" view (the
one with all stories for given component), as MSW loads all these
stories in same iframe/window. Meaning that the last defined MSW mock
will take priority.

So if one story returns a different response for `/assets` endpoint, it
will work when viewed separately (and AFAIK in Chromatic, as it doesn't
care about "Docs" view), but in "Docs" view, all `/assets` calls will
return identical response (the last defined mock).

This is a known limitation as I learned and I haven't found a nice
solution. The one that would work would require us having parameters
being passed to components (and then further to queries and all places
that make the API call) so that finally it will end up added as search
query. The only way for MSW mock to return different response
(regardless of being rendered alone or among other stories) is to have
the URL being different. I am not sure if we are ok and willing to make
such a change in the code to make the "Docs" view work.

### 👀 Preview steps

1. ℹ️ run storybook (`npm run storybook`)
2. Open http://localhost:6006/
3. Go to `DeleteAccountBanner` stories
5. 🟢 Notice that "User Has Assets" story displays "You need to delete or
transfer ownership of all projects…" message
6. 🟢 Notice that "User Has No Assets" story displays "Delete your
account and all your account data." message
7. 🟢 Notice that "User Owns MMO" story displays "You need to transfer
ownership of your organization…" message

---------

Co-authored-by: Kalvis Kalniņš <kalvis.kalnins@kobotoolbox.org>