fix(exports): only try to clean up deployed versions DEV-1053 (#6339)
### 💭 Notes
Developer-only change to only run `repair_file_column_content_in_place`
on deployed versions of an asset. Hopefully this will stop the OOM
errors affecting some exports.
Tested in a python shell on global. There isn't an easy way to
recreate/test the problem locally and this is meant to be a quick patch
so preview steps are left out.