test(backend): mock invalid non-XLS URL import (#6364)
### 💭 Notes
Use a non-xls-string instead of hitting Google so we're not subject to
Google's whims when trying to test what to do with non-xls input.
Updated both v1 and v2 tests to avoid any confusion when eventually v1
is removed (ie to make it obvious that they are redundant and do the
same thing).