feat(dataCollectors): cleanup admin UI DEV-1158 (#6369)
### 📣 Summary
Minor improvements to the Django admin UI for data collectors and data
collector groups.

### 📖 Description
Order data collectors and data collector groups by date created in the
list view. Also change the kobocat url "link" to plain text when in the
edit view for a data collector to make it clear that it's not actually a
link, just an example of the base url that Kobocat will use for this DC.


### 👀 Preview steps

1. ℹ️ have a superuser account
2. In Django admin, add several new data collectors and several new data
collector groups
3. Go to the list view
4. 🟢 Both list views should have the objects listed by recency, with the
most recently created being first
5. Edit a data collector
6. 🟢 "KoboCollect" should be plain text rather than a URL