fix(dataCollectors): remove assets from groups if owners lose permission DEV-1150 (#6357)
### 📣 Summary
Ensure data collectors can no longer access projects if the group owner
has lost permission.


### 📖 Description
Data collector group owners can only add an asset to their group if they
have the manage-asset permission. If that permission is removed, the
asset should no longer be assigned to the group and all the associated
data collector enketo links should be removed.

### Notes
The big problem that needed to be solved with this PR was the situation
where, as part of recalculating permissions, we delete all old ones
before possibly adding them back. To make matters even more complicated,
we do this every time we save an asset.

The approach this PR takes is to refactor the recalculating of
permissions to only delete the old permissions that we know we want to
delete, rather than deleting and re-adding the ones we care about,
except in the case where `return_instead_of_creating` is passed to the
recalculation method. In that case, it's the responsibility of the
caller to handle whatever problems may arise from removing the
manage_asset permission. This approach has a few benefits:
1. no new db calls. this actually reduces the number of db calls because
we aren't re-creating the same permissions that already exist
2. no data collector-specific code outside of the data collector app
3. we can enforce the removal of assets at the model level, no matter if
the permission was removed permissions from the DCG owner via the
expected endpoint(s) or a different path

Other approaches considered:
1. Updating the bulk delete endpoint to call `post_remove_perm` for each
permission removed and using that as the signal instead of
ObjectPermission.post_delete. Rejected because it would add at least one
database call to the endpoint (to get all existing permissions so we
knew which ones we were removing) and because it's endpoint-specific and
not future-proof against other code that deletes permissions.
2. Updating `_recalculate_inherited_perms` to call `post_remove_perm`
for each permission that is *really* removed instead of removed and
added back. Rejected because it would similarly add a database call
3. Updating the custom save() method on ObjectPermission to fire a
custom signal when the object was *really* being removed and not just
temporarily. Rejected because there are cases where
`_recalculate_inherited_perms` will delete a permission and not add it
back, and we would have no way of knowing if the first deletion is
temporary or not.


### 👀 Preview steps

1. ℹ️ have at least 2 accounts and a project
5. Grant manage_asset permissions to user 2
6. In Django admin, create a data collector group with user 2 as the
owner and assign the project to it
7. Add a data collector to the group 
8. Copy the URL link from the data collector's admin page (you either
need to Save and Continue Editing or just reopen it after you save)
9. Verify the URL opens the project
10. Revoke user 2's manage_asset permissions
11. Refresh the data collector URL (removing the x from the URL to make
sure it isn't cached)
12. :red_circle: [on main] survey still opens
13. 🟢 [on PR] error
14. Repeat 2-10 but remove user 2 entirely from the project instead of
just changing permissions