perf(migrations): instructions for faster migration of logger 0049 (#6389)
### 💭 Notes
Use SKIP_HEAVY_MIGRATIONS when updating the instance history table to
allow null values. We don't need to do anything else because the work
for on_delete is done in Python, not SQL, so there's no trigger or
anything that needs to be modified.


### 👀 Preview steps

1. ℹ️ have an account and a project with some submissions
2. Edit a submission to make sure there is stuff in InstanceHistory
3. In a kpi shell, run `./manage migrate 0048 --database kobocat` to go
back a migration
4. In a kpi shell, run `SKIP_HEAVY_MIGRATIONS=True ./manage migrate 0049
--database kobocat`
5. In a postgres shell, run the sql command that is printed
6. Delete the edited submission
7. 🟢 notice that the InstanceHistory objects have instance_id=None
8. In a kpi shell, run `SKIP_HEAVY_MIGRATIONS=True ./manage migrate 0048
--database kobocat` (should be a noop)
9. Locally update the code for InstanceHistory.instance to
`on_delete=models.CASCADE` and remove `null=True` to simulate reverting
the code
10. Edit and delete another submission
11. 🟢 (regression test) notice the related InstanceHistory object(s) are
deleted