fix(data): dynamically add empty validation status dict DEV-1152 (#6388)
### 📣 Summary
Fix error when loading data tables for submissions created before
validation_statuses were added.

### 💭 Notes
Set `_validation_status` to `{}` if it's None on the Instance (which may
happen if instances are bulk updated or created) when saving to mongo,
and, on the other side, set `_validation_status` to `{}` when returning
mongo results if it would otherwise be empty.

### 👀 Preview steps


1. ℹ️ have an account and a project
2. Add submissions to the project
3. In a mongo shell, run `db.instances.updateMany({"_userform_id":
"<username>_<asset_uid>"}, {$unset: {"_validation_status":""}})` . This
will mimic old instances where `_validation_status` isn't present
4. Go to the data table for the project
5. 🔴 [on main] Error
6. 🟢 [on PR] Data table loads successfully