test(xforms): improve xform list API tests DEV-1188 (#6413)
### 💭 Notes
Update XForm list api tests to be more sensitive to changes in the
request flow. Recently we had a few bugs make it into production that
had to do with how the request is handled before it gets to the view.
Since the existing tests were only at the view level, it did not catch
the errors.
While including the whole request processing in test makes it less of a
"unit" test, it makes the tests more likely to catch errors that occur
as a result of the combination of view code and request processing code.
Since the view code relies on information set by django's request
processing, it makes sense to test them together.
There is more consolidation that could be done on these tests as many
are probably redundant (especially testing anonymous access to various
endpoints) but that is outside the scope of this work.


### 👀 Preview steps
Unit-test only, but a good verification is to locally revert the changes
from https://github.com/kobotoolbox/kpi/pull/6412 and
https://github.com/kobotoolbox/kpi/pull/6397, and run the test suite.
Several tests should fail, then succeed when the changes are put back
in.