fix(fromBuilder): accessing non existent attribute for group type DEV-1206 (#6419)
### 📣 Summary
This PR fixes an error when loading form builder containing a group.

### 💭 Notes
The type `group` doesn't contain the attribute `rowType`, which is being
accessed when calculating the `isSupportedByUI`.
The fix just adds an optional chaining, which lets the attribute resolve
to `undefined` and fall into the default clause of being true.

### 👀 Preview steps
1. ℹ️ have an account
2. Create a project with a group. e.g:
[audio_demo.xlsx](https://github.com/user-attachments/files/23193243/audio_demo.xlsx)
3. open the form builder to edit the project
4. 🔴 [on main] notice an error on console and infinite loading
5. 🟢 [on PR] editor opens as expected with no errors