fix(allauth): redirect from signup page DEV-975 (#6426)
### 📣 Summary
Correctly redirect users from the signup page to the landing page
specified in the URL.


### 💭 Notes
Add the redirect url to the login form in a hidden field. Overrides
get_email_confirmation_url to add the redirect url on to the
confirmation link. Add the redirect url in a hidden field on the confirm
form. This will allow django allauth to redirect properly.

### 👀 Preview steps

1. Go `/accounts/signup/?next=/#/account/plan`
2. Create a new user
3. Click the link in the confirmation email
4. Confirm the email
5. 🟢 [on PR] The page automatically lands on the account plans page