refactor(reactQuery): standard error flow for Orval DEV-948 (#6439)
### 💭 Notes

Let's define expected behavior for API error handling on frontend as
follows (see `onErrorDefaultHandler`):
* backend owns error content: keep error messages updated in backend,
don't ignore/override them on frontend.
* frontend owns error display: ideally nicely inline, by default with a
toast using this default handler.
 
With that in mind, this PR introduces a default error handler
`onErrorDefaultHandler` for the react query and refactors all
react-query instances to use it properly. Dealing with frontend error
message overrides will be done seperately as part of DEV-1218.

Out of scope of this PR are non-react-query API calls and the backend
side for them, as both will be a long way to go.

This PR likely displays a few error toasts more than before, but I call
it `refactor` instead of `fix` because AFAIK the error handling so far
was undefined behavior and thus no expected behavior was changed. Let me
know if you disagree.

### 👀 Preview steps

1. ℹ️ have two pending invites to your org
2. open member list in two windows side by side (switching tabs will
make react-query to refetch)
3. remove the invite in one tab
4. change the invite's role in the other tab
6. 🟢 [on PR] notice a nice error toast (for ServerError)
7. invite the user again
8. in devtools set network to offline
4. change the invite's role again
6. 🟢 [on PR] notice a nice error toast (for TypeError), actually two for
both mutation and invalidation