feat(mfa): migrate trench data DEV-854 (#6443)
### 💭 Notes
The trench MFA data is copied to the allauth Authenticator and the
MfaMethodsWrapper models in order to remove the trench dependency. In
order to avoid disabling MFA for users it migrates the user MFA data
from trench when the user logs in. It also includes a long running
migration to migrate users MFA data in the background.