feat(analysis): introduce optional clear button DEV-1194 (#6445)
### 📣 Summary
Modern implementations of radio components don't allow de-selection.
After updating the select one question type in the qualitative analysis
section to mantine, we lost this functionality. To ensure the ability to
de-select select one questions, they will appear with a "clear" button
next to the "delete" button

### Notes
Prerequisite for DEV-1169

### 👀 Preview steps
1. ℹ️ have an account and a project with NLP questions and submissions
2. go to qualitative analysis
3. add a select-one question type
4. 🔴 [on main] notice that there is no clear button
5. 🟢 [on PR] notice the clear button exists and works