build: use latest version of `Formpack` to add `big-image` support DEV-1172 (#6476)
### 📣 Summary
Update KPI to use the latest Formpack release, which adds support for
`big-image` as a valid media type in export workflow.


### 📖 Description
This PR updates the formpack dependency hash in `requirements.in`,
`requirements.txt`, and `dev_requirements.txt` to use the commit from
[formpack#336](https://github.com/kobotoolbox/formpack/pull/336). That
PR introduces support for `big-image` as a valid media type to prevent
translation mismatches in exports.