test(subsequences): fix broken unit tests (#6491)
### Notes
Unit tests only. Skips tests that we eventually want to implement but
don't have implementations for yet. DRF failures are unrelated to PR.
The only substantive difference is in how we add supplements to
duplicated submissions. The old `update_submission_extras` method has
been removed so instead we just create a SubmissionSupplement object
with the correct data.

---------

Co-authored-by: John N. Milner <john@tmoj.net>