feat(subsequences): add model and new endpoints for advanced actions DEV-1229 (#6492)
### 💭 Notes
Add new QuestionAdvancedAction model and associated CRU endpoints. This
PR does not involve actually using the models, though it does include
audit logs for when users hit those endpoints. QuestionAdvancedAction
logs cannot be deleted.
Also includes the automatic migration of the `advanced_features` dict
into corresponding QuestionAdvancedAction objects. For now it does not
change the `advanced_features` dict since we are still using it, but
eventually it will be updated to signal that the data in it has already
been migrated and we should use the associated QuestionAdvancedAction
models instead.
The OpenAPI errors are pre-existing and will be dealt with at the branch
level sometime before merging the full project branch.

### 👀 Preview steps

1. ℹ️ have an account and a project with an audio question 
2. POST to `/api/v2/assets/<asset_uid>/advanced-features/` the following
data:
```
{
     "action": "manual_transcription",
     "question_xpath": <audio question xpath>,
     "params": [{"language": "en"}]
}
```
4. Navigate to `/api/v2/assets/<asset-uid>/advanced-features` in a
browser
5. 🟢 There should be one advanced feature in the list
6. Note the uuid of the action you just created
7. PATCH `/api/v2/assets/<asset_uid>/advanced-features/<action_uuid>/`
with `{"params": ["language": "es"]}`
8. Reload `/api/v2/assets/<asset-uid>/advanced-features`
9. 🟢 [on PR] notice that the params for the action now include both
English and Spanish