refactor(analysis): migrate select one form to mantine DEV-1169 (#6462)
### 💭 Notes
The "best" solution would probably be to refactor all of `radio.tsx` to
mantine, but that creates a bigger problem than this scope. Our
`radio.tsx` (not using mantine) has the ability to de-select options, if
we convert it to use mantine `Radio`, then there will be a regression to
this feature. To offset this for the qualitative analysis project we
introduced a clear button--this has not been done everywhere a
`radio.tsx` is being used. Further investigation into this is needed,
but out of scope of this PR

### 👀 Preview steps
1. ℹ️ have an account and a project with a submission that can open NLP
2. go to qualitative analysis and create a select one question
4. 🔴 [on main] notice that select options are de-selectable
6. 🟢 [on PR] notice that select options are not de-selectable and the
clear button works as intended
8. 🟢 notice that all other functionality is preserved

---------

Co-authored-by: Paulo Amorim <pauloamorimbr@gmail.com>