feat(usage): add LLM requests usage card under feature flag DEV-732 (#6510)
### 💭 Notes
Updates types to get ready for LLM requests stats from backend.
Currently these aren't implemented with any limits/addons, so these are
just placeholders for now. We also don't have any LLM in place, so the
only way to change the usage is manually changing the numbers, which
isn't super useful for now

On the code changes side, I've simplified some of the styling logic with
mantine so we no longer rely on JS to check if there is an add on to
make the relevant cards vertically responsive

### 👀 Preview steps
1. ℹ️ have an account
2. got to addons page
3. purchase an addon such that it will show up in the usage cards
4. go to usage page
6. 🟢 [on PR] notice that the cards are responsive horizontally and
vertically
8. 🟢 notice that the cards a functionally the same as before (no
regressions)