feat(subsequences): show supplemental columns in data table DEV-1387 (#6523)
### 📣 Summary
Add supplemental NLP columns to data table.

### 📖 Description
This is just for adding the columns to the data table. They may not be
populated correctly. If an NLP action is enabled, there will be a column
for it, even if there are presently no responses.

### 💭 Notes
Using analysis_form_json to avoid having to make changes on the frontend
even though it's not a very descriptive name. Does not add QA questions,
those will come later.
Removed some of the response fields from the old analysis_from_json
since they don't seem to be used. It's possible in QA or when dealing
with exports they will turn out to be used but we can always add them
back.

### 👀 Preview steps

1. ℹ️ have an account and a project with an audio question with at least
one response
2. In a python shell, enable all NLP actions by running
```
        asset = Asset.objects.get(uid={uid})
        for action in [
            Action.MANUAL_TRANSLATION,
            Action.MANUAL_TRANSCRIPTION,
            Action.AUTOMATIC_GOOGLE_TRANSLATION,
            Action.AUTOMATIC_GOOGLE_TRANSCRIPTION,
        ]:
            language = 'en' if 'transcription' in action else 'es'
            QuestionAdvancedFeature.objects.create(
                question_xpath={xpath},
                action=action,
                params=[{'language': language}],
                asset=asset,
            )
```
Note: this will enable English manual/automatic transcripts and Spanish
automatic/manual translations
3. Navigate to the data table
4. 🟢 [on PR] Notice there are columns for English transcript and Spanish
translation for the relevant question