feat(subsequences): migrate old SubmissionExtras to SubmissionSupplemental DEV-934 (#6422)
### 💭 Notes
Fill out the method for converting old SubmissionExtra content dicts to
the new format expected by SubmissionSupplemental for translations and
transcripts.
This code makes numerous assumptions to fill in information that is not
present in the old structure but required in the new:

1. If old[xpath]['transcript']['value'] ==
old[xpath]['googlets']['value']and the language codes are the same, we
assume the most recent transcript was automatically generated
2. If, for any revision in old[xpath]['transcript']['revisions'],
revision['value'] is the same as old[xpath]['googlets']['value'] and the
language codes match, we assume that revision was automatically
generated. If multiple match, we assume they were all automatically
generated. This should be pretty rare but it's possible
3. 1-2 also apply to transcriptions
4. old[xpath]['transcript']['dateModified'] will be assumed to be the
creation date of the most recent revision (ie whatever is in
old[xpath]['transcript']['value']). The same goes for translations
5. All uuids are newly generated
6. All old transcriptions/translations have status=complete with a
_dateAccepted of now() (whenever the code is running)
7. To determine the dependency of any old translation, whether automated
or manual:

* If we know the source language, look for the most recent transcript in
that language that was created before the translation.
* If there is none, take the most recent transcription in that language
* If there are no transcriptions in the source language, take the most
recent transcript
* If we don't know the source language, take the most recent transcript
that was created before the translation
* If there is none, take the most recent transcript

8. We can ignore any badly formatted revisions/transcripts/translations
9. Most recent revisions will be first in the version array