feat(subsequences): allow hiding of QA questions DEV-935 (#6550)
### 💭 Notes
Automatically set `{"options": {"deleted": True}` if a user attempts to
delete a QA question via a PATCH request.
As part of the `update_params` method of the `qual` action, we replace
the entire `options` dict within each question with whatever comes in
the request. This is because the only thing I know of in the `options`
dict is the `deleted` param, which we want to be able to change on the
front end. If it turns out we need the `options` dict for more, we may
have to write a more sophisticated way of merging old and new options.

In the future we may want to replace the options dict with just a
boolean for `hidden,` but this way minimizes the additional work the
frontend will have to do to work with the new API.


### 👀 Preview steps

1. ℹ️ have an account and a project with an audio question and at least
one submission
2. Enable QA with `curl -X POST -H 'Authorization: Token <your token>'
http://kf.kobo.local/api/v2/assets/<asset_uid>/advanced-features/ --json
'{"question_xpath": <xpath>, "action": <qual>, "params": [{"type":
"qualInteger", "label": {"_default": "How many?"},
"uuid":"11111111-1111-1111-1111-1111111111"}]}'`
3. Note the uid of the action that comes back in the response
4. Try to "remove" the question with `curl -X PATCH -H 'Authorization:
Token <your token>'
http://kf.kobo.local/api/v2/assets/<asset_uid>/advanced-features/<feature_uid>/
--json '{"params": []}'`
5. 🟢 [on PR] notice that the "How many?" question is still in the
response, with `{"options": {"hidden": True}}`

---------

Co-authored-by: Olivier Léger <olivierleger@gmail.com>