Run pip-compile.sh (raw output)
I had isolated the sorting changes from the parent commit
(721cf48e6d11e4ad0c51a67bcfcf965e9319947b) to produce a nicer diff
for code review.

But here is the actual output from uv.

It sorts the git+https dependencies to be among the rest of the
dependencies (instead of a separate block at the top).

Future runs of pip-compile.sh would match this.